+++ lib/pro_image_editor.dart
@@ -71,6 +71,7 @@ export 'core/constants/editor_style_constants.dart';
 export 'core/utils/image_converter.dart';
 export '/shared/utils/parser/int_parser.dart';
 export '/shared/utils/parser/double_parser.dart';
+export '/shared/utils/parser/offset_parser.dart';
 export '/shared/utils/parser/size_parser.dart';
 export '/core/models/editor_configs/utils/editor_safe_area.dart';